+++ trunk/openvibe-applications/designer/trunc/src/ovdCInterfacedScenario.cpp
@@ -1022,16 +1022,16 @@ void CInterfacedScenario::addCommentCB(int x, int y)
 		::GtkAdjustment* l_pVAdjustment=gtk_scrolled_window_get_vadjustment(GTK_SCROLLED_WINDOW(l_pScrolledWindow));
 
 #if defined OVD_OS_Linux && !defined OVD_OS_MacOS
-		x=gtk_adjustment_get_value(l_pHAdjustment)+gtk_adjustment_get_page_size(l_pHAdjustment)/2;
-		y=gtk_adjustment_get_value(l_pVAdjustment)+gtk_adjustment_get_page_size(l_pVAdjustment)/2;
+		x=(int)(gtk_adjustment_get_value(l_pHAdjustment)+gtk_adjustment_get_page_size(l_pHAdjustment)/2);
+		y=(int)(gtk_adjustment_get_value(l_pVAdjustment)+gtk_adjustment_get_page_size(l_pVAdjustment)/2);
 #elif defined OVD_OS_Windows
 		gint wx,wy;
 		::gdk_window_get_size(gtk_widget_get_parent(GTK_WIDGET(m_pScenarioDrawingArea))->window, &wx, &wy);
-		x=gtk_adjustment_get_value(l_pHAdjustment)+wx/2;
-		y=gtk_adjustment_get_value(l_pVAdjustment)+wy/2;
+		x=(int)(gtk_adjustment_get_value(l_pHAdjustment)+wx/2);
+		y=(int)(gtk_adjustment_get_value(l_pVAdjustment)+wy/2);
 #else
-		x=gtk_adjustment_get_value(l_pHAdjustment)+32;
-		y=gtk_adjustment_get_value(l_pVAdjustment)+32;
+		x=(int)(gtk_adjustment_get_value(l_pHAdjustment)+32);
+		y=(int)(gtk_adjustment_get_value(l_pVAdjustment)+32);
 #endif
 	}
